feat: refactor code + algorithm selection
