+++ src/controllers/openai.controllers.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { GenResProps, createResPrompt, sendClassicPrompt } from "../openai/prompts";
+import { GenResProps, createResPrompt, sendClassicPrompt } from "../execution/prompts";
 import { genGraphChatGpt } from "../openai/generators";
 
 export async function genResource(req: Request<any, any, GenResProps>, res: Response){
