+++ src/types/PolyglotFlow.ts
@@ -1,11 +1,16 @@
 import { PolyglotEdge, PolyglotNode } from ".";
 
+export type PolyglotExecutionData = {
+    algo: string
+}
+
 export type PolyglotFlowInfo = {
     _id: string;
     title: string;
     author: string;
     description: string;
     tags: string[];
+    execution: PolyglotExecutionData;
 }
 
 export type PolyglotFlow = PolyglotFlowInfo & {
