+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from "./PolyglotEdge"
 export * from "./PolyglotNode"
 export * from "./PolyglotFlow"
+export * from "./PolyglotConcept"
